requirements.txt
# Core
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
email-validator==2.1.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Cache
redis==5.0.4

# Scheduler
apscheduler==3.10.4

# Email
aiosmtplib==3.0.1
jinja2==3.1.4

# Crypto / SSH
cryptography==42.0.5

# Proxmox (optionnel si PROXMOX_ENABLED=true)
proxmoxer==2.0.1
urllib3==2.2.1
